feat: restructure app settings and distortion profile components,
