+++ CustomHeadsetGUI/README.md
@@ -12,3 +12,18 @@
 2. install [node.js](https://nodejs.org/) & [rust](https://www.rust-lang.org/tools/install)
 3. run `npm i` in `CustomHeadsetGUI`
 4. run `npm run build` in `CustomHeadsetGUI`
+5. output to `..\output\CustomHeadsetGUI`
+
+## Project Structure
+
+```
+CustomeHeadsetGUI
+├───public //angular static files
+├───src //angular sources
+│    ├────app
+│    │     ├───dialogs  //dialog , all components opened using MatDialog in here
+│    │     ├───pages    //app pages,all tabs main components
+│    │     └───services //services,all injectable service
+│    └────locale //i18n strings
+└───src-tauri //tauri sources
+```
