+++ CustomHeadsetGUI/src/app/pages/basic-settings/basic-settings.component.html
