+++ CustomHeadsetGUI/src/app/pages/basic-settings/basic-settings.component.scss
@@ -1,4 +1,4 @@
-@use '../../simple-form.scss' as sf;
+@use '../../../simple-form.scss' as sf;
 
 :host{
     @include sf.formStyle();
