+++ CustomHeadsetGUI/src/app/pages/distortion-profile/distortion-profile.component.scss
